Time: 16 ms (87.65%), Space: 22.4 MB (92.52%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -29,7 +29,7 @@ public:
                 if(node->right) q.push(node->right) ;
                 avg += node->val ;
             }
-            ans.push_back(avg/size) ;
+            ans.push_back((double)avg/size) ;
         }
         return ans ;
     }
